Fix uniform class weights
Class weights should be floats. Other options implicitely convert to
float when dividing by the class counts.

+++ src/lesion_gnn/datasets/base.py
@@ -85,7 +85,7 @@ class BaseDataset(InMemoryDataset):
         counts = self.classes_counts
         match mode:
             case ClassWeights.UNIFORM:
-                return torch.ones_like(counts)
+                return torch.ones_like(counts).float()
             case ClassWeights.INVERSE:
                 return 1 / counts
             case ClassWeights.QUADRATIC_INVERSE:
